Update text shown after submitting a data request
Trying to add a clearer note to users that data requests take a couple weeks to be reviewed (discussed with PS). Removed some of the extraneous text bits.

Note that the bolded text would be confusing for "immediate access" datasets, but there aren't any datasets of this type (yet). A fancier idea might be to have this text customizable per dataset, but that's another conversation.